v0.1.1.0
Updated to support RimWorld version 1.3. Added patches for new Bulk recipes added by dismarzero. Included a patch to add a research prerequisite (Synthread Research) for the Plasteel Thread recipe since Plasteel Thread is useless without the Synthread research. This prevents the recipe from cluttering the list when it is not yet useful.